+++ scripts/test_train.py
@@ -7,7 +7,7 @@ import fire
 basedir = Path(__file__).parent
 
 
-def main(lang="mecab", pretrained=""):
+def main(lang="ja_mecab_torch", pretrained=""):
     if not pretrained:
         pretrained = str((basedir / "../tests/fixtures/bert").absolute())
     cfg = {
@@ -26,7 +26,7 @@ def main(lang="mecab", pretrained=""):
 
 def matrix():
     fixturesdir = basedir / "../tests/fixtures"
-    langs = ["mecab", "juman", "sentencepiece"]
+    langs = ["ja_mecab_torch", "ja_juman_torch", "sentencepiece_torch"]
     pretrained_dirs = [
         str((fixturesdir / "bert").absolute()),
         str((fixturesdir / "xlnet").absolute()),