+++ tests/conftest.py
@@ -85,8 +85,17 @@ def device(request):
     return torch.device("cuda")
 
 
-@pytest.fixture(scope="session", params=["mecab", "juman", "sentencepiece"])
+TORCH_LANGS = ["ja_mecab_torch", "ja_juman_torch", "sentencepiece_torch"]
+ALL_LANGS = ["ja_mecab", "ja_juman", "sentencepiece"] + TORCH_LANGS
+
+
+@pytest.fixture(scope="session", params=ALL_LANGS)
 def lang(request):
+    return request.param
+
+
+@pytest.fixture(scope="session", params=TORCH_LANGS)
+def torch_lang(request):
     return request.param