+++ tests/test_spacy_factory.py
@@ -1,7 +1,0 @@
-import pytest
-import spacy
-
-
-@pytest.mark.parametrize("lang", ["mecab"])
-def test_spacy_factory(lang):
-    spacy.blank(lang)